+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## [Unreleased]
 
-## Added
+### Added
 
 - [TD-4986] Configurable data catalog bucket views based on note and metadata
   fields
 - [TD-5787] Add Elastic Search Boost option in templates
 - [TD-5840] Fix color of result in implementations search
+- [TD-5593] Add related concept in quality implementations list and implementation download
 - [TD-5808] Configured TdCluster
 
 ## [5.9.0] 2023-06-20